Renames app package
